+++ src/main/java/baseball/domain/User.java
@@ -4,14 +4,14 @@ import baseball.view.InputView;
 
 public class User {
 
-    private int userNumber;
+    private String userNumber;
 
     public User() {
 
     }
 
     public void setUserNumber() {
-        this.userNumber = InputView.inputUserNumber();
+        this.userNumber = InputView.inputConsole();
     }
 
     public int getUserNumber() {
