+++ src/main/java/baseball/view/InputView.java
@@ -1,24 +1,18 @@
 package baseball.view;
 
-import baseball.controller.InputValidator;
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputView {
-    public static int inputUserNumber() {
+    public static String inputConsole() {
         OutputView.printInputUserNumberMessage();
         String value = Console.readLine();
-        if (InputValidator.inputUserNumberValidator(value)) {
-            return Integer.parseInt(value);
-        }
-        return inputUserNumber();
+        validateIsEmpty(value);
+        return value;
     }
 
-    public static int inputRestartOrNotNumber() {
-        OutputView.printRestartOrNotMessage();
-        String value = Console.readLine();
-        if (InputValidator.inputRestartOrNotNumberValidator(value)) {
-            return Integer.parseInt(value);
+    private static void validateIsEmpty(String value) {
+        if (value.isEmpty()) {
+            throw new IllegalArgumentException("값을 입력해주세요.");
         }
-        return inputRestartOrNotNumber();
     }
 }
